+++ packages/thumbprint-react/CHANGELOG.md
@@ -5,6 +5,8 @@
 ### Changed
 
 -   [Patch] Refactor the internals of the `Image` component.
+-   [Minor] Add `onClick` prop to `ServiceCard`.
+-   [Minor] Add `target` prop to `ServiceCard`.
 
 ## 2.0.0 - 2019-06-03